Make rummager app ID and router paths configurable for whitehall.
The whitehall app needs to run a separate instance of Rummager
(whitehall-rummager in alphagov-deployment) and this needs to
register paths with a different prefix with the router to discriminate
them from the citizen search paths. Also it needs to register these paths
against a different backend URL, i.e. the one for the whitehall instance of
Rummager.