Avoid "no such file to load -- delsolr" error.
This was the error I was seeing when serving the app locally using
either rackup or pow. It's caused because we're trying to require
"delsolr" before adding it to the load path. I've chosen to prevent this
by telling Bundler to require all the libraries we need before requiring
`app.rb`. This introduced another problem because I don't have RACK_ENV
set on my local machine. It turns out that `env.rb` already takes care
of setting this to a default value of "development" so I'm requiring
that before bundler.

My understanding is that this was working OK on production (under
Passenger) because Passenger does some magic if you have a Gemfile in
your app.