Don't register whitehall search with the router
In whitehall, the search is always proxied through the main whitehall
app and the public never accesses whitehall-search. Therefore there's no
need to register whitehall-search with the router.

Furthermore the /government/search url should be served by the whitehall
app and the fact that it was routed to whitehall-search broke things.