Missing section description doesn't 404 browse page
This fixes an issue where a missing section description, e.g.
_business.html for the 'business' section, would cause a page to 404
(rather than 500) when this view file was missing.

This case has been wrapped in a rescue block in case this mistake occurs
in the future.

Also added in a _business.html description file so that we remember to
fill it in when we go live.