Disable SimpleCov by default.
I was getting annoyed by the coverage directories appearing in
various directories within my project, depending on where I ran the
tests from.  I've set the USE_SIMPLECOV environment variable in Jenkins
so that the reports continue to be generated over there.

I realise that this might be controversial so feel free to change it
back if it upsets your workflow.