Allow search results to be filtered by format.
This is primarily driven by our desire to show only specialist guides in
the search results on /specialist.  Specialist guides are currently
stored in the whitehall app and are indexed with a format of
'specialist_guide'.  By allowing us to pass the optional format_filter
parameter to rummager, we can restrict our search to just those
documents of a certain type.

I've updated the whitehall-rummager schema in the puppet repo so that it
indexes the format field, the same as it is in mainstream rummager.
This is required for the filtering to work at all.

This commit is a work in progress.  We'll also need to update the
autocomplete so that it can also filter by a given format.  If we don't
do that then the autocomplete will continue to show results from all of
IG (obviously only in the whitehall app), while the actual results from
/specialist will only contain specialist guides.