Use Document as source of truth for humanized format name.
The logic to choose the format name to display was split between
two helper methods, each of which gave slightly different answers.

Inside Government also needs access to this information, so rather
than duplicating the logic again, the presentation format and
humanized format name are now returned in the JSON results, to be
used to build the Inside Government search results page.