External links have external view metadata
External links (currently referred to as recommended links due to the
repo of the same name that exists) should have slightly different view
classes and attributes to take advantage of the design work that we
already have in place.

These are purely cosmetic changes and the tests make sure that these
attributes are correctly set in the event that results show external
(recommended) links in the output.