Avoid reading boost file for every indexing call.
This was what was giving us terrible performance (~1 document/sec)
through single registration, as the documents were being submitted
individually and triggering a full CSV file read each time. The file is
only going to change on deployment, as it's currently implemented, so
there's no obvious risk to reading it all in on first use.

We're likely to revisit document boosting soon, so this may be removed
in its current form, but this was the simplest change that would give us
the performance we want.