Ensure we only deal with elasticsearch indexes
I was trying to create the indexes on preview but the task failed
because we have both solr and elasticsearch backends on preview.  The
`rummager_environment` task was only ignoring backends that had a type
of 'none' which meant we were trying to instantiate an
`ElasticsearchAdminWrapper` for the solr backend.