Group tests into separate rake tasks - units, functional & integration.
I was seeing an intermittent Nokogiri segmentation fault when running
the tests locally. With all the tests run in a single rake task it was
really hard to see what was going on. In particular it feels weird that
we would load all the integration-related stuff when only running the
unit tests, especially as it seems as if the unit tests were running
*after* the other tests.