Make dependency on `ElasticsearchAdminWrapper` explicit.
This class is used within `IntegrationTest`, but there is no
corresponding require statement. The tests were all passing by chance,
because an individual test was requiring the relevant file before the
class was first used, but without this extra require it was impossible
to run some of the tests individually e.g. `ElasticsearchBrowsingTest`.