Allow the use of a "subsubsection" key.
This is needed for detailed guidance in order to e.g. show individual
breadcrumbs alongside search results.

While it would have been nicer and more "scalable" to combine the
"section", "subsection" & "subsubsection" keys into a single key with
e.g. an array of values, this would require significant changes in many
places. The solution in this commit seems like the most consistent and
least worst option for the moment.