Fix a bug in deletion when links are full URLs.
Because the link wasn't being Lucene-escaped in the query, Lucene's
query parser was getting confused by the colon in the `link` field.

Now we escape this using Lucene query escaping, but pass it into the
`params` argument to RestClient rather than passing it through
`CGI.escape` to avoid confusing double-escaping.