Default to `all_documents` returning everything.
Instead of requiring a maximum number of documents, just default to a
number significantly greater than the number of documents indexed.

There's a chance this will come back to bite us, but by the time the
search index grows large enough to hit this limit, we will have had to
change the way this code works anyway, because pulling out all the
documents into memory will have become impractical.