Horrific hack to fix some indices not restoring.
In Jenkins and on development machines with ~20 or more shards, the
`test_should_return_symbol_if_index_exists` test fails (intermittently
on Jenkins; consistently on the development machines) when waiting for
the index to reopen, as some shards fail to restore and are left
unassigned so the cluster remains in a "red" state.

Pausing briefly before closing the index fixes this, but I don't have a
clear idea why.