Time decay boosting for content with a public timestamp
Implemented using a custom score as in this blog post:
http://jontai.me/blog/2013/01/advanced-scoring-in-elasticsearch/

Changed `now` to time() and have not included a general boot of +1 as
this is only applied to documents where the public_timestamp exisits.

Also had to increase the boosts on the Inside Gov formats to compensate
for the time boosts.

This is for: https://www.pivotaltracker.com/story/show/42780733