Remove the concept of a secondary search index.
This was a temporary hack to mix together mainstream results and
detailed guidance: since these are now presented as separate
result streams (and if we do join them back together it'll be in
a very different way), we can remove secondary search.

The main search endpoint now just becomes a special case of the
`/backend_name/search` endpoint, where the backend name is blank.
We also don't need explicit references to the primary search in the
controller code any more.