Add document_series to mappings
We want to index document_series so that we can show them in the search
results. This will help people to recognise whether the result is
relevant to them.

We need to add it to the mappings file to prevent this field from being
analyzed (ie. stemmed).

This will need to be deployed with deploy:migrations at a quiet time.
When migrating the mappings rummager will build a new index and migrate
the data over to it. If a document gets added to the original index
during migration then the migration will fail.

https://www.pivotaltracker.com/story/show/50503119