Health check support for multiple backend types
Changed the health check script to be a command line script rather than
a rake task as it's easier to accept and parse command line options.

I've used a gem called 'slop' for command line parsing:
  https://github.com/injekt/slop

Added the html and json backends to support the old ways of running
health check. This should mean we can delete the govuk-search-health
repo.