Add search clause to boost promoted items
Promoted items will have the promoted_for field set to a list of
keywords. This is an analyzed field so promotional terms will be stemmed
etc.

I've changed the search query to add an outer bool query with two should
clauses:
  i) the previous bool query containing the main search part for the
phrase (with minimum_should_match) and should clauses for boosting
  ii) a clause which matches only against the promoted_for field

Any documents matching promoted_for will immediately be boosted to the
top of the results. All other aspects of the search query remain
unchanged.