Fix auth credentials in the health_check script.
From the looks of it, this bit of functionality didn't properly survive the
transition from `govuk_search_health` to its current home in Rummager: one
side of the script was treating it as a string; the other trying to splat it
into `basic_auth`. We can use a custom value type to handle parsing at one end
and splatting at the other.

As part of this, also fix misnamed attributes in the two search clients, which
were breaking the auth reporting.