Specify an explicit encoding for Aspell.
Without this, Aspell tries to fall back on determining a locale from the
environment: if this fails (as it can do when running within Upstart), it
returns a value of "none" for the encoding, which blows up when ffi-aspell
tries to use it.

This is unfortunately tricky to test without delving deep into the Aspell
library, so a strongly-worded comment and the knowledge that Preview will show
up any failures will have to do for now.