"wales office" wasn't returning the org page
The Wales Office organisation page[0] should be returned for "wales office".

The query "wales office" was expanded to "wales cymru cymraeg gymraeg office".
With this number of terms, the minimum should match filter requires at least 3
terms to be in the document, but the org page only has "wales" and "office", so was
excluded from the results.

I wasn't able to find any examples of results that this change makes worse, but
it does fix this case.

  * 0 - https://www.gov.uk/government/organisations/wales-office