Index documents asynchronously.
This lets Rummager handle the case where the search engine is not
available or a request times out, rather than falling over with an
error and leaving the requesting application to handle the failure.

This is a first pass at getting something that works, so it's not very
pretty and doesn't have good test coverage.

For the purposes of testing, for now, we're disabling job queueing and
making all jobs happen synchronously. This is in lieu of making the
tests aware of how the queues work.