Base the queue's namespace on the environment.
More than anything else, this is to avoid any overspill between
development and testing. The tests should all be unable to write to
the queue, but just in case, this shields us from accidentally trying
to post test data and getting IndexNotFound errors.

Hat tip to @jamiecobbett for the idea.