Unify config use across Rake, workers and Sinatra.
We have three ways of running code from Rummager:

  * From the Rakefile, for test runs and index manipulation jobs
  * From the `sidekiq` script, through `bootstrap_worker.rb`
  * From a Rack server, through `config.ru` and `app.rb`

It's a faff keeping the load paths and requirements in line across all
three of these, which isn't worth the purity of keeping the Sinatra code
out of the non-Sinatra environments.

As a side-effect, we can now remove an `any_instance` stubbing in the
integration test helper, since worker code now accesses the searcn
configuration through Sinatra settings like everything else.