Comment specific synonym tweaks.
These synonyms are intended to manipulate the order of results for
particular queries, and as such are likely to be prone to breaking when
the content or the search algorithm changes. Documenting them now helps
us keep track of what they're for, and gives us a better idea of what
new features would be useful in Rummager to tweak result ordering.