Use mr-sparkles instead of shotgun
These two gems provide reloading of code in development. Shotgun does this by
booting the entire app for every request. This is slow. mr-sparkles on the
other hand, only reloads when files change, so is much faster. Additionally,
shotgun has an open bug in that it doesn't log in development, which is
annoying.

Also this means using unicorn in development, which should be good for dev-prod
parity.

I had to fork the gem and make changes to it. This is because it uses the
Listen library which uses filesystem events if they are available. Filesystem
events are available on our VMs, but don't work for files changed on the host
machine because they're mounted by the VM over NFS.