Add some backtrace information to failed jobs.
Because the failed jobs mail queue is our only source of error reporting, we
should have information on where in the code the exception happened.
Empirically, 12 lines is about what we need to get from RestClient errors out
to the top level of the app, without going out into the Sinatra gem and the
various levels of Rack middleware.