Added new stats release announcement fields
A new content type is being added to whitehall called
StatisticalReleaseAnnouncement. This will be used to pre-announce
forthcoming statistical releases. We require two extra fields in the
search index to allow us to list these in search results:

expected_release_timestamp - a timestamp representing the expected
release of the stats for sorting purposes (so we can show the
soonest first)

expected_release_text - a textual version of the expected release
date to be displayed on the site, as the timestamp will sometimes
only be an approximation, e.g. "April 2015".