Fix deletion with elasticsearch 0.90+.
Document deletion uses the delete_by_query API in elasticsearch; the way
in which that expects queries to be escaped has been changed in
elasticsearch 0.90 compared to 0.20, such that documents were not being
deleted by the existing code in elasticsearch 0.90+.  The simplest fix
is just to use the request body to send the query using the query DSL,
since this avoids needing to do any escaping.  Avoiding doing escaping
is better anyway.