Allow searching across multiple indexes.
This adds support for SearchServer.index to be given a comma separated
list of index names instead of a single index name.  Elasticsearch
understands this syntax for referencing multiple indexes.

This paves the way for us to perform a search across multiple indexes in
elasticsearch, rather than having to join the results in Ruby.