Add Elasticsearch.Index.raw_search().
This performs a search directly in elasticsearch given a payload.  It
abstracts a little common code from the "search" and "advanced_search"
methods, but I mainly want to do this so that I can call this directly
from the new UnifiedSearcher.

Why start breaking the abstraction that Elasticsearch-specific stuff
should live in "elasticsearch"?  Because it's a false abstraction;
rummager only supports elasticsearch as a backend and there's no clean
api for the elasticsearch module; elasticsearch specific idioms and
schema handling code permeate most of the rummager code.  Ultimately we
should get replace most of the code in the elasticsearch directory by
using the official elasticsearch client.