Add script to dump links.
This script simply iterates through all indexes, dumping the contents of
the "link" field, and the index name, as one entry per line.  This is
intended to be used to check the contents of an index, both now (to
resolve https://www.pivotaltracker.com/story/show/67319636 ) and
whenever we are concerned that there may be documents missing or present
that shouldn't be.

It should take no more than a few seconds to run.