Keep metadata separate for UnifiedSearchPresenter
Instead of merging the metadata fields (eg, _score, _index) with the
rest of the fields for passing input to the UnifiedSearchPresenter, put
them all into a :_metadata member.  This avoids risk of them colliding
with other fields, and makes it easier to write tests that check all the
concrete fields have the correct values.