Rename multi_index_helper to multi_index_test.
This is in line with the convention that the filename should match the
name of the class defined in it.  The file will be included when tests
are run, which may waste a little time in test runs, but on the other
hand if a developer in future adds a test to the file it won't silently
fail to be run.