Update health check to support new search endpoints
This changes the health check component so that the JSON client now
supports the "unified" and "combined" search endpoints provided by
Rummager.

It drops support for the old-style Content API /search.json endpoint
which was not used, and drops support for checking an individual index.
This is because indexes are no longer queried on an individual basis -
rather, weightings and filters are applied at the new endpoints.