Use old query structure to improve ranking
Copy the query structure used for the old endpoint to improve the
ranking produced.  This ranking isn't great, but is significantly better
than the pure elasticsearch ranking, and we're going to start by just
making minor tweaks to this until we have the ability to use features
based on popularity and click through as a base of our ranking.