Bias search results away from government content
The large number of long government documents overwhelms the other
content somewhat, so multiply the scores of government documents by a
constant to reduce them.  The constant chosen here is 0.4, which is the
result of running the health check with various different constants, and
choosing the value which gave the best balance of government versus
other documents.