Add popularity field to documents when indexing
This field is populated from the page-traffic index, which in turn is
populated from a dump file produced by the python scripts in the
search-analytics repository.

The popularity is currently defined here as 1.0 / the rank of the page
in terms of traffic.  ie, Our highest traffic page gets 1.0, the next
highest gets 0.5, the next highest gets 0.333.  This is a nice and
stable measure.