Separate index names into content and auxiliary.
The page-traffic index is an example of an auxiliary index; it doesn't
hold content, but is used outside the main query (in this case, at
indexing time). We'll shortly be adding another such index to hold
best-bets.

Sitemaps should be made only from content indexes.  The search server
needs to know all the allowed names for its validation of index names
when opening indexes.

Conflicts:
	elasticsearch.yml