Configure page_traffic in the test environment
This renames "page-traffic" to "page_traffic", so that we can adopt the
convention of appending "_test" to the index in the test environment.

This also adds configuration for "page_traffic_test", and changes
Elasticsearch::Index to automatically infer the traffic index based on
matching against the auxiliary indexes that are present.