Explicitly define the rank_14 field in the schema
Already, "rank_14" field is defined as a dynamic field, however when no
documents are present in the index, querying on this field will result
in Elasticsearch returning a mapping error. It is possible this scenario
could occur if we were to deploy this change without seeding data into
the popularity index.

Defining the field again separately is enough to fix this.