Add disable_best_bets debug option
This adds a generic debug parameter, which accepts a comma separated
list of options. There are plenty of options we can add in future, such
as providing access to elasticsearches explain API so we can understand
the weighting better.  For now though, just add an option
"disable_best_bets" which disables the use of best bets for query
ranking.