Remove old promoted links code
This is replaced by the new best-bets implementation.  Best bets for
"jobs" and "visa" are populated on production in the new system, and we
want to retire the promoted link for "budget 2014", so the new
implementation covers all the result adjustments made by this code.

This implements https://www.pivotaltracker.com/story/show/71295084