Report errors in workers to errbit
Notify errbit only after sidekiq has given up retrying the message
(though sidekiq is currently configured not to retry).  Airbrake is
given an exception containing details of the failure; it would be nicer
to give it the original exception so that traceback, etc, was
meaningful, but I can't see a way to do that easily and this is better
than nothing.