Add disable_synonyms debug option
If debug=disable_synonyms is added to the query parameters, this
disables use of the synonym list by using an analyzer which doesn't
include synonyms in its list of actions.  This is useful for working out
whether synonyms are helping to fix results for particular queries, or
whether they're causing problems.