Add tags and specialist_sectors to schema
The specialist sectors field is not indexed, because it's going to be
merged into the tags field.  It does need to appear in the "_source"
JSON of the document, though, so that amend actions work.  (Amend works
by fetching the _source field for the document, updating the fields
listed in the amendments, and then indexing it again.)

The tags field is just a non-analyzed field, holding a list of values,
which will be used for filtering and faceting.