Allow extra options to be supplied for faceting
We want to allow callers to request that a set of example documents are
returned when faceting.  As a first step towards this, this commit
changes the facet parameters to accept a set of additional
comma-separated facet parameters; within each facet parameter, keys and
values are separated by colons.  For example:

    facet_organisations=10,examples:5,example_fields:link:title,example_scope:global

The parameters will actually be used by subsequent commits.