Add updated_at field to core doctype schema
To be able to support finders from rummager we need to be able to sort
documents by their updated_at timestamp. We don't ever want to search on
the field, just sort on it, so index it but don't include it in `_all`.

Note that whilst this field is common across all doctypes, it's not
mandatory. Until everything is using the new content store some things
will have null updated_at timestamps.