Analyze the specialist_sectors field
We were originally not analyzing the specialist_sectors field because we
had a plan to move to storing this in an internal tags field.  However,
we don't currently have any searches which require a single internal
tags field, so this is probably a misstep for now.  Changing the field
to be analyzed will make it available for searches after we next run a
deploy:migrations.

Also, don't allow use of the tags field in searches, since it shouldn't
be visible externally.