Add a parse_query_string helper
This uses Rack's KeySpaceConstrainedParams to get protection from
callers making requests with very large numbers of keys, but otherwise uses much
simpler logic than the normal query parsing.  Parameters are always
returned as lists of values, and square bracket special-syntax handling
is restricted to stripping off trailing "[]" from parameter names (so
that we support ruby clients which add that automatically for values
which are lists).