Fake a slug for sector documents
Specialist sector documents don't come from whitehall and so they don't
have a slug in search.  Panopticon constructs the link field from the
slug (held in panopticon) by adding a '/' prefix, so rather than do the
work now to add a slug and reindex the sectors, we construct a slug
field for the cached sector results in SpecialistSectorRegistry by
removing this '/'.

If the slug field is added in future, this hack can be removed, but it
shouldn't break anything if it isn't.

Fixes https://www.pivotaltracker.com/story/show/74843014