Use CGI::parse to parse query parameters
This is nicer (less code) than using our hand-rolled parser, though
CGI::parse's code is pretty similar to what we were doing.

This loses the use of Rack's KeySpaceConstrainedParams to limit the
number of parameters, but this is mainly useful for preventing DOSes on
post requests, and we only expose GET endpoints to the public, so this
should be fine.